copy to clipboard, move to Stage
