+++ examples/worldview/package.json
@@ -15,6 +15,7 @@
     "d3-request": "^1.0.6",
     "global": "^4.3.0",
     "hubble.gl": "^1.1.0",
+    "react-copy-to-clipboard": "^5.0.3",
     "react-helmet": "^6.1.0",
     "react-map-gl": "^5.2.10",
     "react-markdown": "^4.2.2",
